+++ src/app/providers/index.ts
@@ -1,1 +1,2 @@
+export * from './chat.provider';
 export * from './feedback.provider';